ZenohTest/pyproject.toml
[project]
name = "zenohtest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eclipse-zenoh>=1.7.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
]
